pyproject.toml
[project]
name = "bigquery"
version = "0.1.0"
description = "bigquery python lib"
authors = [
    {name = "ppabam", email = "tom@diginori.com"},
]
dependencies = ["google-cloud-bigquery>=3.31.0"]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true


[dependency-groups]
air = [
    "apache-airflow==2.10.2",
]
